modifications of the german text
the german text appearing on the website was modified and extended.
there is still the problem of the last index.html version modifies by andrii terliuk:
if the generated diagram will be stored, the corresponding thumbnail diagram shows the frame only, not the diagram. 
but it is stored in dir plots
this problem has to be solved.